+++ mayan/apps/tags/tests/mixins.py
@@ -1,0 +1,42 @@
+from __future__ import unicode_literals
+
+from documents.permissions import permission_document_view
+from documents.tests import GenericDocumentViewTestCase
+
+from ..models import Tag
+from ..permissions import (
+    permission_tag_attach, permission_tag_create, permission_tag_delete,
+    permission_tag_edit, permission_tag_remove, permission_tag_view
+)
+
+from .literals import (
+    TEST_TAG_COLOR, TEST_TAG_COLOR_EDITED, TEST_TAG_LABEL,
+    TEST_TAG_LABEL_EDITED
+)
+
+
+class TagTestMixin(object):
+    def _create_tag(self):
+        self.tag = Tag.objects.create(
+            color=TEST_TAG_COLOR, label=TEST_TAG_LABEL
+        )
+
+    def _request_tag_create_view(self):
+        return self.post(
+            viewname='tags:tag_create', data={
+                'label': TEST_TAG_LABEL,
+                'color': TEST_TAG_COLOR
+            }
+        )
+
+    def _request_tag_delete_view(self):
+        return self.post(
+           viewname='tags:tag_delete', args=(self.tag.pk,)
+        )
+
+    def _request_tag_edit_view(self):
+        return self.post(
+            viewname='tags:tag_edit', args=(self.tag.pk,), data={
+                'label': TEST_TAG_LABEL_EDITED, 'color': TEST_TAG_COLOR_EDITED
+            }
+        )